feat: add read operation for qq

+++ src/quizQuestions/api/quizQuestions.controller.ts
@@ -2,28 +2,37 @@ import {
   Body,
   Controller,
   Delete,
+  Get,
   HttpCode,
   HttpStatus,
+  NotFoundException,
   Param,
   ParseIntPipe,
   Post,
   Put,
+  Query,
   UseGuards,
 } from '@nestjs/common';
 import { BasicAuthGuard } from '../../features/auth/guards/basic-auth.guard';
 import {
   PublishQuestionUpdateModel,
   QuizQuestionsCreateModel,
+  SortingQueryParamsForQuiz,
 } from './models/input/quiz-questions.input.model';
 import { CommandBus } from '@nestjs/cqrs';
 import { CreateQuizQuestionCommand } from '../application/use-cases/create-quiz-question-use-case';
 import { DeleteQuizQuestionCommand } from '../application/use-cases/delete-quiz-question-use-case';
 import { UpdateQuizQuestionCommand } from '../application/use-cases/update-quiz-question-use-case';
 import { PublishQuestionCommand } from '../application/use-cases/publish-question-use-case';
+import { QuizQuestionsQueryRepository } from './infrastructure/quiz-questions.query.repository';
+import { FindAllQuestionsCommand } from '../application/use-cases/get-all-quiz-questions-use-case';
 
 @Controller('sa/quiz/questions')
 export class QuizQuestionsController {
-  constructor(private commandBus: CommandBus) {}
+  constructor(
+    private commandBus: CommandBus,
+    private quizQueryRepo: QuizQuestionsQueryRepository,
+  ) {}
   @UseGuards(BasicAuthGuard)
   @Post()
   @HttpCode(HttpStatus.CREATED)
@@ -42,6 +51,10 @@ export class QuizQuestionsController {
   async deleteQuestion(
     @Param('id', ParseIntPipe) id: number,
   ): Promise<boolean> {
+    const isQuestionExist = await this.quizQueryRepo.findQuestion(id);
+    if (!isQuestionExist) {
+      throw new NotFoundException();
+    }
     return await this.commandBus.execute(new DeleteQuizQuestionCommand(id));
   }
 
@@ -52,6 +65,10 @@ export class QuizQuestionsController {
     @Param('id', ParseIntPipe) id: number,
     @Body() quizUpdateModel: QuizQuestionsCreateModel,
   ): Promise<boolean> {
+    const isQuestionExist = await this.quizQueryRepo.findQuestion(id);
+    if (!isQuestionExist) {
+      throw new NotFoundException();
+    }
     return await this.commandBus.execute(
       new UpdateQuizQuestionCommand(quizUpdateModel, id),
     );
@@ -64,8 +81,23 @@ export class QuizQuestionsController {
     @Param('id', ParseIntPipe) id: number,
     @Body() publishQuestionUpdateModel: PublishQuestionUpdateModel,
   ): Promise<boolean> {
+    const isQuestionExist = await this.quizQueryRepo.findQuestion(id);
+    if (!isQuestionExist) {
+      throw new NotFoundException();
+    }
     return await this.commandBus.execute(
       new PublishQuestionCommand(id, publishQuestionUpdateModel),
+    );
+  }
+
+  @UseGuards(BasicAuthGuard)
+  @Get()
+  @HttpCode(HttpStatus.OK)
+  async getAllQuestions(
+    @Query() sortingQueryParamsForQuiz: SortingQueryParamsForQuiz,
+  ) {
+    return await this.commandBus.execute(
+      new FindAllQuestionsCommand(sortingQueryParamsForQuiz),
     );
   }
 }
